fix llm tool schema shaping for near ai
