+++ src/llm/mod.rs
@@ -36,6 +36,7 @@ pub mod runtime;
 pub mod session;
 pub mod smart_routing;
 mod token_refreshing;
+pub(crate) mod tool_schema;
 pub mod transcription;
 
 #[cfg(test)]
